handwritten-digit-recognition/requirements.txt
numpy
Pillow
streamlit
matplotlib
seaborn
scikit-learn
tensorflow
opencv-python-headless
streamlit-drawable-canvas